Draw accurate machine and keypad
The vending machine and keypad have been redesigned to match the real
life machine. The smaller keypad on the machine itself is now just a
scale()d down version of the full-size one, meaning it will show hovered
buttons and text on the screen without the need for a second set of
positions and scales.