Add items and rewrite animStages
All of the animStages have been rewritten as vendStages (to avoid name
collisions), up until the item needs to zoom into view and open the more
detailed UI view. The images for the items have also been added,
although a non-selectable placeholder needs to be put in the empty
spaces.